[Doc] Edit for grammar and style (#985)
* Edit for grammar and style

In the opening paragraph, it would be helpful to provide some overall scenario and the prerequisites you expect readers to have completed before they start here. "This tutorial assumes you have already..." Add a link to Install DGL topic perhaps and any other framework or IDE or even specialized knowledge. With this context, you help readers to succeed by setting expectations.

Better to group all tutorials (this, PageRank with, Batched Graph, Working with) at the same level in the left navigation rail and all as subsections of DGL Basics.

* Update tutorials/basics/2_basics.py